requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
babel==2.5.3              # via django-phonenumber-field
click==6.7                # via pip-tools
django-phonenumber-field==2.0.0
django==2.0.5
first==2.0.1              # via pip-tools
phonenumberslite==8.9.5   # via django-phonenumber-field
pip-tools==2.0.2
psycopg2-binary==2.7.4
pytz==2018.4              # via babel, django
selenium==3.12.0
six==1.11.0               # via pip-tools
